refactor(desktop): modularize oauth and extract user creation
