+++ apps/desktop/src-tauri/src/oauth/mod.rs
@@ -1,0 +1,3 @@
+pub mod session_store;
+
+pub use session_store::OAuthSessionStore;
